requirements.txt
streamlit
pandas
plotly
psycopg2-binary
beautifulsoup4
requests